update dockerfiles and exploration OP
