+++ requirements.txt
@@ -1,5 +1,5 @@
 numpy
-setuptools-scm
+setuptools_scm
 scikit-learn
 pydflow
 tensorflow==2.6.2
