+++ rid/common/mol.py
@@ -151,11 +151,3 @@ def slice_dump(
         )
     else:
         raise RuntimeError("Unknown Style for Slicing Trajectory.")
-    
-def final_dump(
-    dump: str,
-    selected_idx,
-    output_format:str
-):
-    traj = dpdata.System(dump, fmt="lammps/dump")
-    traj.to('lammps/lmp', output_format, frame_idx=selected_idx)
